# packages/api/src/openapi/openapi-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Stagehand API",
    "version": "1.0.0",
    "description": "A comprehensive API for managing security governance across organizational hierarchies",
    "contact": {
      "name": "Stagehand Development Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie authentication"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
      "Team": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "CreateTeamRequest": {
        "type": "object",
        "required": [
          "name",
          "companyId"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Team name"
          },
          "description": {
            "type": "string",
            "description": "Team description"
          },
          "companyId": {
            "type": "string",
            "description": "Parent company ID"
          }
        }
      },
...
      "UpdateTeamRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Team name"
          },
          "description": {
            "type": "string",
            "description": "Team description"
          }
        }
      },
...
      "CreateRelationshipRequest": {
        "type": "object",
        "required": [
          "sourceProjectId",
          "targetProjectId",
          "type"
        ],
        "properties": {
          "sourceProjectId": {
            "type": "string",
            "description": "Source project ID"
          },
          "targetProjectId": {
            "type": "string",
            "description": "Target project ID"
          },
          "type": {
            "type": "string",
            "enum": [
              "DEPENDS_ON",
              "USES",
              "INTEGRATES_WITH",
              "EXTENDS"
            ],
            "description": "Type of relationship"
          },
          "description": {
            "type": "string",
            "description": "Optional description of the relationship"
          }
        }
      },
      "ProjectRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Relationship ID"
          },
          "sourceProjectId": {
            "type": "string",
            "description": "Source project ID (the project that depends on another)"
          },
          "targetProjectId": {
            "type": "string",
            "description": "Target project ID (the project being depended upon)"
          },
          "type": {
            "type": "string",
            "enum": [
              "DEPENDS_ON",
              "USES",
              "INTEGRATES_WITH",
              "EXTENDS"
            ],
            "description": "Type of relationship between projects"
          },
          "description": {
            "type": "string",
            "nullable": true,
            "description": "Optional description of the relationship"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UpdateRelationshipRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "DEPENDS_ON",
              "USES",
              "INTEGRATES_WITH",
              "EXTENDS"
            ],
            "description": "Updated relationship type"
          },
          "description": {
            "type": "string",
            "description": "Updated description of the relationship"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/teams/{id}": {
      "put": {
        "summary": "Update team details",
        "description": "Updates team information such as name and description",
        "tags": [
          "Teams"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Team successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to update this team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete team",
        "description": "Permanently deletes a team and all associated data",
        "tags": [
          "Teams"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Team ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Team successfully deleted"
          },
          "403": {
            "description": "Not authorized to delete this team",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/teams": {
      "get": {
        "summary": "List all teams user has access to",
        "description": "Returns all teams that the authenticated user has access to view",
        "tags": [
          "Teams"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of teams",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new team",
        "description": "Creates a new team within a company",
        "tags": [
          "Teams"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Team successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Team"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to create team in this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Company not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/relationships": {
      "get": {
        "summary": "Get project relationships",
        "description": "Retrieves all relationships for projects within a specific company",
        "tags": [
          "Relationships"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "companyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID to get project relationships for"
          }
        ],
        "responses": {
          "200": {
            "description": "List of project relationships",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectRelationship"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Company ID is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to view relationships for this company",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create project relationship",
        "description": "Creates a new dependency relationship between two projects",
        "tags": [
          "Relationships"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRelationshipRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Relationship successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRelationship"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request data or cannot create self-referencing relationship",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to create relationships between these projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "One or both projects not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Relationship already exists between these projects",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/relationships/{id}": {
      "put": {
        "summary": "Update project relationship",
        "description": "Updates an existing project relationship's type or description",
        "tags": [
          "Relationships"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Relationship ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRelationshipRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Relationship successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRelationship"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized to update this relationship",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Relationship not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete project relationship",
        "description": "Removes a project relationship",
        "tags": [
          "Relationships"
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Relationship ID"
          }
        ],
        "responses": {
          "204": {
            "description": "Relationship successfully deleted"
          },
          "403": {
            "description": "Not authorized to delete this relationship",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Relationship not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "security": [
    {
      "cookieAuth": []
    }
  ]
}